refactor: move each feature on server to separate directory
