+++ package.json
@@ -22,6 +22,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "express-promise-router": "^4.1.1",
     "express-session": "^1.17.3",
     "fluent-ffmpeg": "^2.1.2",
     "pg": "^8.11.3",
